backend/requirements.txt
aiofiles==25.1.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
bcrypt==4.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-swagger-dark==0.0.7
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
minio==7.2.16
passlib==1.7.4
pillow==11.3.0
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
reportlab==4.4.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
